run postgres tests in CI?

+++ tests/conftest.py
@@ -57,11 +57,14 @@ async def ledger():
         await ledger.load_used_proofs()
         await ledger.init_keysets()
 
-    db_file = "test_data/mint/test.sqlite3"
-    if os.path.exists(db_file):
-        os.remove(db_file)
+    # clear sqlite database
+    if not settings.mint_database.startswith("postgres"):
+        db_file = os.path.join(settings.mint_database, "test.sqlite3")
+        if os.path.exists(db_file):
+            os.remove(db_file)
+
     ledger = Ledger(
-        db=Database("test", "test_data/mint"),
+        db=Database("test", settings.mint_database),
         seed=settings.mint_private_key,
         derivation_path=settings.mint_derivation_path,
         lightning=FakeWallet(),
